+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.xx-dev
 Release date: ??.??.????
 
+* HUB-784 - Improving accessibility on degree programme and other education provider selections.
 
 ## 1.53
 Release date: 27.10.2020